[Process] Add support for streams as input
